Add NOTION_TOKEN existence check + split out Notion getData function
